Move joystick code to the keyboard level.
